Move utils to correct folders
