--- src/api/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './jsonrpc';
-export * from './validators';
-export * from './store';
-export * from './hashing';
