+++ src/app/components/Core/TimeElapsed.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
 import { useInterval } from '@app/hooks';
-import { formatTimeDifference } from '@utils';
+import { formatTimeDifference } from '@common/utils';
 
 const TimeElapsed = ({ value }: { value: number }) => {
   const [timeElapsed, setTimeElapsed] = useState(formatTimeDifference(value));
