+++ src/app/utils/userInput.ts
@@ -1,8 +1,8 @@
 import { hexlify } from '@ethersproject/bytes';
 import { formatEther, formatUnits, parseEther, parseUnits } from '@ethersproject/units';
 
+import { addHexPrefix, bigify } from '@common/utils';
 import type { TransactionRequest } from '@types';
-import { addHexPrefix, bigify } from '@utils';
 
 export type HumanReadableTx = ReturnType<typeof toHumanReadable>;
