+++ src/common/utils/hex.test.ts
@@ -1,4 +1,4 @@
-import { addHexPrefix } from '@utils/hexPrefix';
+import { addHexPrefix } from '@common/utils/hex';
 
 describe('addHexPrefix', () => {
   it('adds hex prefix if no prefix exist', () => {
@@ -7,7 +7,7 @@ describe('addHexPrefix', () => {
     );
   });
 
-  it('adds doesnt add hex prefix the string is already prefixed', () => {
+  it("doesn't add hex prefix the string is already prefixed", () => {
     expect(addHexPrefix('0x4bbeEB066eD09B7AEd07bF39EEe0460DFa261520')).toBe(
       '0x4bbeEB066eD09B7AEd07bF39EEe0460DFa261520'
     );
