+++ src/common/utils/truncate.ts
