+++ src/types/bigish.ts
@@ -1,0 +1,3 @@
+import type BigNumber from 'bignumber.js';
+
+export type Bigish = BigNumber;
