+++ src/utils/index.ts
@@ -1,8 +1,5 @@
-export * from './generateUUID';
-export { toChecksumAddress } from './toChecksumAddress';
-export * from './truncate';
-export * from './tx';
-export * from './hexPrefix';
-export * from './json';
-export * from './date';
-export * from './bigify';
+export * from './encryption';
+export * from './hashing';
+export * from './jsonrpc';
+export * from './store';
+export * from './validators';
